# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreateVercelSandboxRequestImage = typeof CreateVercelSandboxRequestImage[keyof typeof CreateVercelSandboxRequestImage];


export const CreateVercelSandboxRequestImage = {
  'vercel/sandbox/node:22': 'vercel/sandbox/node:22',
  'vercel/sandbox/python:313': 'vercel/sandbox/python:3.13',
} as const;

export interface CreateVercelSandboxRequest {
  /**
     * @minLength 3
     * @maxLength 48
     * @pattern ^[a-z0-9][a-z0-9-]*$
     */
  name: string;
  image: CreateVercelSandboxRequestImage;
}

export type VercelSandboxInfoImage = typeof VercelSandboxInfoImage[keyof typeof VercelSandboxInfoImage];


export const VercelSandboxInfoImage = {
  'vercel/sandbox/node:22': 'vercel/sandbox/node:22',
  'vercel/sandbox/python:313': 'vercel/sandbox/python:3.13',
} as const;

export interface VercelSandboxInfo {
  name: string;
  image: VercelSandboxInfoImage;
  timeout: '1h';
}

export interface ErrorResponse {
  message: string;
}

export interface TerminalTicketInput {
  sudo: boolean;
}

export interface TerminalTicket {
  ticket: string;
}