# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: vercel-sandboxes
    description: Authenticated Vercel Sandbox operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /vercel-sandboxes:
    post:
      operationId: createVercelSandbox
      tags: [vercel-sandboxes]
      summary: Create a Vercel Sandbox
      description: Creates a time-limited Vercel Sandbox for an authenticated mobile client.
      security:
        - TerminalGateway: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVercelSandboxRequest"
      responses:
        "201":
          description: Sandbox created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VercelSandboxInfo"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid gateway token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Vercel Sandbox CLI could not create the sandbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: API server credentials are not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vercel-sandboxes/{name}:
    delete:
      operationId: stopVercelSandbox
      tags: [vercel-sandboxes]
      summary: Stop a Vercel Sandbox
      description: Stops a sandbox while retaining its filesystem snapshot.
      security:
        - TerminalGateway: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 48
            pattern: "^[a-z0-9][a-z0-9-]*$"
      responses:
        "204":
          description: Sandbox stopped
        "400":
          description: Invalid sandbox name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid gateway token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Vercel Sandbox CLI could not stop the sandbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: API server credentials are not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vercel-sandboxes/{name}/terminal-ticket:
    post:
      operationId: createVercelSandboxTerminalTicket
      tags: [vercel-sandboxes]
      summary: Create a one-time Vercel Sandbox terminal ticket
      description: Issues a short-lived single-use ticket for a WebSocket terminal connection.
      security:
        - TerminalGateway: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 48
            pattern: "^[a-z0-9][a-z0-9-]*$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminalTicketInput"
      responses:
        "200":
          description: Single-use terminal ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminalTicket"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid gateway token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: API server credentials are not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    TerminalGateway:
      type: http
      scheme: bearer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CreateVercelSandboxRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 48
          pattern: "^[a-z0-9][a-z0-9-]*$"
        image:
          type: string
          enum:
            - vercel/sandbox/node:22
            - vercel/sandbox/python:3.13
      required:
        - name
        - image
    VercelSandboxInfo:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        image:
          type: string
          enum:
            - vercel/sandbox/node:22
            - vercel/sandbox/python:3.13
        timeout:
          type: string
          const: 1h
      required:
        - name
        - image
        - timeout
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        message:
          type: string
      required:
        - message
    TerminalTicketInput:
      type: object
      additionalProperties: false
      properties:
        sudo:
          type: boolean
          default: false
      required:
        - sudo
    TerminalTicket:
      type: object
      additionalProperties: false
      properties:
        ticket:
          type: string
      required:
        - ticket